+++ src/DemoFile.Game.Deadlock/DeadlockDemoParser.cs
@@ -1,4 +1,5 @@
-﻿using DemoFile.Sdk;
+﻿using DemoFile.Game.Deadlock;
+using DemoFile.Sdk;
 
 namespace DemoFile;