+++ src/DemoFile.Game.Deadlock/Source1GameEvents.Autogen.cs
@@ -5,7 +5,7 @@
 using System.Text.Json.Serialization;
 using DemoFile.Sdk;
 
-namespace DemoFile;
+namespace DemoFile.Game.Deadlock;
 
 public partial class Source1GameEvents
 {