+++ src/DemoFile.Test/Integration/HttpBroadcastReaderTest.cs
@@ -1,0 +1,6 @@
+﻿namespace DemoFile.Test.Integration;
+
+public class HttpBroadcastReaderTest
+{
+    
+}