+++ src/DemoFile.Test/MockHttpMessageHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace DemoFile.Test;
+
+public class MockHttpMessageHandler
+{
+    
+}